requirements.txt
flask==3.0.0
twilio==8.12.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.15.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0